gui-main/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robot-sim-engine"
version = "0.7.0"
description = "Pure-math robotic arm kinematics and trajectory planning simulator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
gui = [
  "PySide6>=6.5",
  "pyqtgraph>=0.13",
  "pyvista>=0.43",
  "pyvistaqt>=0.11",
  "pytest-qt>=4.4",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.10",
  "pre-commit>=3.8",
]

[project.entry-points."robot_sim.plugins"]

[tool.setuptools.packages.find]
where = ["src"]


[tool.setuptools.package-data]
robot_sim = [
  "resources/configs/*.yaml",
  "resources/configs/profiles/*.yaml",
  "resources/configs/robots/*.yaml",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]


[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
follow_imports = "skip"
files = [
  "src/robot_sim/core",
  "src/robot_sim/domain",
  "src/robot_sim/model/trajectory.py",
  "src/robot_sim/model/task_snapshot.py",
  "src/robot_sim/model/version_catalog.py",
  "src/robot_sim/model/app_config.py",
  "src/robot_sim/model/solver_config.py",
  "src/robot_sim/app/version.py",
  "src/robot_sim/app/version_catalog.py",
  "src/robot_sim/app/runtime_paths.py",
  "src/robot_sim/application/workers/task_events.py",
  "src/robot_sim/application/services/manifest_builder.py",
  "src/robot_sim/presentation/error_boundary.py",
  "src/robot_sim/presentation/view_contracts.py",
  "src/robot_sim/presentation/main_window_ui.py",
  "src/robot_sim/presentation/coordinators",
  "src/robot_sim/presentation/facades.py",
]

[tool.coverage.run]
source = ["src/robot_sim"]
branch = true


[tool.coverage.report]
fail_under = 80
show_missing = true
